SendAsset component
